mips2: change text-addressing scheme
Originally, I was going to do a lot of work to keep *pc relative to
*progbase when I could, but I ran into an issue with jalr and jr, which
use the value in a register… if it's a register *I* set, then it's
likely already relative to *progbase. It could be some type of
jump-table calculation, or who knows what.

It is, in the end, simpler to keep *pc absolute and in the text-segment,
meaning it is always ≥ TEXT_START. The only time I have to dereference
*pc is when reading an instruction. So *then* I can subtract TEXT_START
and add *progbase to find the right address in my data memory:

    *(*pc - TEXT_START + *progbase)

In some comments I have written Mem[addr] rather than *addr, but they
mean the same thing.