mips2: extract pieces of instruction earlier
This saves on instructions by a large amount (~1800 to ~900), simply by
prematurely loading all or almost all relevant pieces.

If these were subroutines, repetition would not have been an issue. But
because I like macros, the expansions result in large amounts of
duplicate code.

A further optimization which I hope not to need would be to make the
largest macro-bodies subroutines and add macro-wrappers that move
registers into arguments and return values into registers; I estimate
the savings here to be minimal (each macro incurs at least a few moves
and a jal, while saving the result of a body, so is probably ~3 or fewer
instructions saved for the most complex macros). Worse, now that the
macros are invoked at most twice, there is very little to gain here.